requirements.txt
fastapi
uvicorn
requests
langchain
langchain-ollama
python-dotenv